+++ frontend/src/app/main/ui/workspace/tokens/management/forms/generic_form.cljs
@@ -10,9 +10,7 @@
    [app.common.data :as d]
    [app.common.files.tokens :as cfo]
    [app.common.schema :as sm]
-   #_[app.common.types.token :as cto]
    [app.common.types.tokens-lib :as ctob]
-   [app.common.uuid :as uuid]
    [app.main.constants :refer [max-input-length]]
    [app.main.data.helpers :as dh]
    [app.main.data.modal :as modal]
@@ -37,12 +35,6 @@
    [cuerdas.core :as str]
    [rumext.v2 :as mf]))
 
-#_(defn- token-value-error-fn
-  [{:keys [value]}]
-  (when (or (str/empty? value)
-            (str/blank? value))
-    (tr "workspace.tokens.empty-input")))
-
 (defn get-value-for-validator
   [active-tab value value-subfield form-type]
 
@@ -58,29 +50,6 @@
       value)
 
     value))
-
-#_(defn- default-make-schema
-  [tokens-tree _]
-  (sm/schema
-   [:and
-    [:map
-     [:name
-      [:and
-       [:string {:min 1 :max 255 :error/fn #(str (:value %) (tr "workspace.tokens.token-name-length-validation-error"))}]
-       (sm/update-properties cto/schema:token-name assoc :error/fn #(str (:value %) (tr "workspace.tokens.token-name-validation-error")))
-       [:fn {:error/fn #(tr "workspace.tokens.token-name-duplication-validation-error" (:value %))}
-        #(not (ctob/token-name-path-exists? % tokens-tree))]]]
-
-     [:value [::sm/text {:error/fn token-value-error-fn}]]
-
-     [:description {:optional true}
-      [:string {:max 2048 :error/fn #(tr "errors.field-max-length" 2048)}]]]
-
-    [:fn {:error/field :value
-          :error/fn #(tr "workspace.tokens.self-reference")}
-     (fn [{:keys [name value]}]
-       (when (and name value)
-         (not (cto/token-value-self-reference? name value))))]]))
 
 (mf/defc form*
   [{:keys [token
@@ -98,7 +67,7 @@
            input-value-placeholder] :as props}]
 
   (let [make-schema     (or make-schema #(-> (cfo/make-token-schema % token-type)
-                                             (sm/dissoc-key :id)))  ;; TODO this does not work because the schema is no longer a :map but a :multi
+                                             (sm/dissoc-key :id)))
         input-component (or input-component token.controls/input*)
         validate-token  (or validator default-validate-token)
 
@@ -134,13 +103,11 @@
 
         initial
         (mf/with-memo [token]
-          (-> (or initial
-              {:id (uuid/next)    ;; TODO it should not be necessary if the sm/dissoc-key :id worked correctly
-               :type token-type
+          (or initial
+              {:type token-type
                :name (:name token "")
                :value (:value token "")
-               :description (:description token "")})
-              (d/tap-r #(prn "initial" %))))
+               :description (:description token "")}))
 
         form
         (fm/use-form :schema schema